Beanstalk HTTP client v0.1.0 Beanstalk.Changeset

Manage your commit histories on your Repositories.

See: http://api.beanstalkapp.com/changeset.html

Functions

all()

Find all Changesets

all(page, per_page, order_field \\ "time", order \\ "DESC")

Find all Changesets (with Pagination)

Args:

page (integer) — page number for pagination;

per_page (integer) — number of elements per page (default 15, maximum 30);

order_field (string) — what column to use for ordering (default is time);

order (string) — order direction. Should be either ASC or DESC (default is DESC).

differences(repository_id, revision)

Find Changeset Diffs

This method will return an array of differences for files of the changeset. Each difference is represented as an array of path, status and diff in Unified Diff format.

Args:

repository_id

revision

find(repository_id, revision)

Find Changeset

Note that revision number is used instead of unique ID, therefore you have to specify repository_id.

Args:

repository_id - repository to find changeset on.

revision - Subversion only. Revision number.

repository(repository_id)

Find all Changesets for Repository

Args:

repository_id - repository to find changesets for.

repository(repository_id, page, per_page, order_field \\ "time", order \\ "DESC")

Find all Changesets for Repository (with Pagination)

Args:

repository_id - repository to find changesets for.

page (integer) — page number for pagination;

per_page (integer) — number of elements per page (default 15, maximum 30);

order_field (string) — what column to use for ordering (default is time);

order (string) — order direction. Should be either ASC or DESC (default is DESC).