Beanstalk HTTP client v0.1.0 Beanstalk.Release

Manage releases for your Repositories.

See: http://api.beanstalkapp.com/release.html

Functions

all()

Returns a list of all releases for the Beanstalk account.

all(page, per_page)

Returns a paginated list of releases within your Beanstalk account.

Set the page number and per page.

Args:

page - page number for pagination.

per_page - number of elements per page (default 30, maximum 50).

create(repository_id, server_environment_id, comment, revision, deploy_from_scratch \\ false, do_not_notify \\ false)

Creating a release means requesting a deployment.

If you already have some releases in state “new” or “retry”, Beanstalk will delete them after creating a new one.

Args:

repository_id - repository to create the release on.

server_environment_id - server environment to use.

comment - comment for the release.

revision - specific commit to release.

deploy_from_scratch - deploy from the first revision rather then doing incremental deployment.

do_not_notify - do not trigger email notification for this deployment.

repository(repository_id)

Returns a list of releases for the given repository.

Args:

repository_id - repository to find releases for.

repository(repository_id, page, per_page)

Returns a paginated list of releases for the given repository.

Args:

repository_id - repository to find releases for.

page - page number for pagination.

per_page - number of elements per page (default 30, maximum 50).

retry(repository_id, release_id)

Use this method to retry a failed deployment.

Args:

repository_id - repository to retry the release on.

release_id - release to retry.