Beanstalk HTTP client v0.1.0 Beanstalk.User

Manage user accounts within Beanstalk.

See: http://api.beanstalkapp.com/user.html

Functions

all()

Returns a full list of users within your Beanstalk account.

all(page, per_page)

Returns a paginated list of users within your Beanstalk account.

Set the page number and per page.

Args:

page - page number for pagination.

per_page - number of elements per page (default 30, maximum 50).

create(login, email, name, password, timezone \\ "", admin \\ false)

Create new user.

Admin privileges required for this API method.

Args:

login - Username. Unique per Account.

email - Email address. Unique per account.

name - Full name.

password - Required on create.

timezone - Optional.

admin - Boolean value. Optional. Default is false.

current()

Returns currently logged in user (that is: the user you configured for module).

delete(user_id)

Delete the given user.

Admin privileges required for this API method.

You can not delete account owner.

Args:

user_id - The id attribute from the user object (not account_id).

find(user_id)

Returns a single user matching the given user_id.

Args:

user_id - The id attribute from the user object (not account_id).

update(user_id, login, email, name, password, timezone \\ "", admin \\ false)

Update the given user.

Admin privileges required for this API method.

Args:

user_id - The id attribute from the user object (not account_id).

login - Username. Unique per Account.

email - Email address. Unique per account.

name - Full name.

password - Required on create.

timezone - Optional.

admin - Boolean value. Optional. Default is false.